Moved the code to accept a follow from user controler to protocol.ts for future use in branch controller
