+++ src/controllers/user.ts
@@ -1,5 +1,6 @@
 import * as utils from "../utils";
 import * as model from "../model";
+import * as protocol from "../protocol";
 const request = require("request");
 
 export async function handleUserInboxPost(url: string[], query: any, req: any, res: any, body: string, cookies: any){
@@ -11,45 +12,16 @@ export async function handleUserInboxPost(url: string[], query: any, req: any, r
         let user: model.User | undefined = await model.getUserByName(userName);
         
         if (user) {
-            await model.createFollow(streamObject.actor, utils.urlForPath('user/' + user.name));
+            protocol.handleFollow(
+                streamObject,
+                utils.urlForUser(user),
+                utils.urlForUser(user) + "#main-key",
+                user.privateKey
+            );
             
             res.statusCode = 201;
             res.end();
             
-            let remoteDomain = streamObject.actor.split("/")[2];
-            
-            let date = new Date().toUTCString();
-            
-            let stringToSign = `date: ${date}`;
-            let signedString = utils.signString(user.privateKey, stringToSign);
-            let header       = `keyId="${utils.urlForPath("user/"+user.name)}#main-key",algorithm="rsa-sha256",headers="date",signature="${signedString}"`;
-            
-            let body = JSON.stringify({
-                    "@context": [
-                        "https://www.w3.org/ns/activitystreams",
-                        "https://w3id.org/security/v1"
-                    ],
-                    
-                    id: utils.urlForPath(`user/${userName}/followaccept/${Math.random()}`),
-                    type: "Accept",
-                    actor: utils.urlForPath("user/" + user.name),
-                    
-                    object: streamObject
-                });
-            
-            let options = {
-                url:  streamObject.actor + "/inbox", // this won't work for some server. Will fix later
-                headers: {
-                    Host      : remoteDomain,
-                    Date      : date,
-                    Signature : header,
-                },
-                body: body
-            };
-            
-            request.post(options, (err: any, resp: any, body: string) => {
-                utils.log("Post to remote instance follow accept answer", err, resp, body);
-            });
         } else {
             res.statusCode = 404;
             res.end("Invalid user");
